usercenter修改信息,球队列表,详情.球员详情

+++ apps/team/logics.py
@@ -10,34 +10,68 @@ from apps.team.models import Team,Player
 from django.contrib.auth.models import User  
 import utils.files.logics as fileLogics
 
+TEAM_DATA = {
+    'name':'',
+    'point':'0',  #  得分
+    'shot':'0/0',  # 投篮
+    'three':'0/0', #三分
+    'rebound':'0', # 篮板
+    'steal':'0', # 抢断
+    'assist':'0', # 助攻
+    'turnover':'0', # 失误
+    'block':'0', # 盖帽
+    'players':[],
+    'game':"0",
+    "win":"0"
+}
+
 # Team do something
 @transaction.atomic
 def saveTeam(user,postData):
-    teams = Team.objects.filter(user=user,status=1)
-    if teams.length > 0:
-        team = teams[0]
+    teams = Team.objects.filter(manager=user,status=1)
+    if len(teams) > 0:
+        return False
     else:
         team = Team()
         team.id_code = fileLogics.getIdCode("TEAM")
         team.manager = user
         user.status = 'manager'
         team.name = postData['name']
-        profile = postData['name']
-        logo = postData['logo']
-    school = postData['school']
-    desc = postData['desc']
-    os.mkdir(os.path.join(settings.TEAM_PROFILE_DIR,team.name))
-    with open(settings.TEAM_PROFILE_DIR+"/"+team.name+"/profile","wb+") as fp:
-        fp.write("")
-    team.save()
-    user.save()
-    fileLogics.setIdCode("TEAM",int(team.id_code)+1)
-    return True;
+        team.profile = team.id_code
+        # logo = postData['logo']
+        team.school = postData['school']
+        team.desc = postData['desc']
+        os.mkdir(os.path.join(settings.TEAM_PROFILE_DIR,team.id_code))
+        with open(settings.TEAM_PROFILE_DIR+"/"+team.id_code+"/profile","w+") as fp:
+            fp.write(json.dumps(TEAM_DATA))
+        team.save()
+        if not user.player is None:
+            user.player.team = team
+            user.player.save() 
+        user.save()
+        fileLogics.setIdCode("TEAM",int(team.id_code)+1)
+        return True;
+
+@transaction.atomic
+def editTeam(user,postData):
+    teams = Team.objects.filter(manager=user,status=1)
+    if len(teams) == 0:
+        return False
+    else:
+        team = teams[0]
+        team.name = postData['name']
+        team.profile = postData['name']
+        team.school = postData['school']
+        team.desc = postData['desc']
+        team.save()
+        return True;
+
 
 @transaction.atomic
 def changeLogo(team,logo):
     oldLogo = team.logo
-    os.remove(os.path.join(settings.UPLOADED_DIR,oldLogo))
+    if not oldLogo is None:
+        os.remove(os.path.join(settings.UPLOADED_DIR,oldLogo))
     team.logo = logo
     team.save()
     return True
@@ -61,16 +95,21 @@ def disbandTeam(team):
 # Player do something
 @transaction.atomic
 def savePlayer(user,postData):
-    if user.player is None:
-        player = Player()
-        player.id_code = fileLogics.getIdCode("PLAYER")
-        player.user = user
-        user.status = 'player'
-        with open(settings.PLAYER_PROFILE_DIR+"/"+id_code,"wb+") as fp:
-            fp.write("")
-        player.profile = id_code
-    else:
-        player = user.player
+    player = Player()
+    player.id_code = fileLogics.getIdCode("PLAYER")
+    player.user = user
+    user.status = 'player'
+    with open(settings.PLAYER_PROFILE_DIR+"/"+player.id_code,"w+") as fp:
+        fp.write("")
+    player.profile = player.id_code
+    height = postData['height']
+    if height[-2:] != 'cm':
+        height = height + 'cm'
+    weight = postData['weight']
+    if weight[-2:] != 'kg':
+        weight = weight + 'kg'
+    player.height = height
+    player.weight = weight
     player.position = postData['position']
     player.desc = postData['desc']
     player.save()
@@ -79,8 +118,30 @@ def savePlayer(user,postData):
     return True
 
 @transaction.atomic
-def changeNum(player,number):
+def editPlayer(user,postData):
+    player = user.player
+    # with open(settings.PLAYER_PROFILE_DIR+"/"+player.id_code,"w+") as fp:
+    #     fp.write("")
+    height = postData['height']
+    if height[-2:] != 'cm':
+        height = height + 'cm'
+    weight = postData['weight']
+    if weight[-2:] != 'kg':
+        weight = weight + 'kg'
+    player.height = height
+    player.weight = weight
+    if player.team is None:
+        player.position = postData['position']
+    player.desc = postData['desc']
+    player.school = postData['school']
+    player.save()
+    return True
+
+
+@transaction.atomic
+def changeNumAndPos(player,number,position):
     player.number = number
+    player.position = position
     player.save()
     return True
 
@@ -97,4 +158,3 @@ def leaveTeam(player):
     player.number = null
     player.save();
     return True
-
